Added JsonIgnore annotations

+++ src/main/java/com/atipera/task/model/ErrorSchema.java
@@ -1,4 +1,4 @@
 package com.atipera.task.model;
 
-public record ErrorSchema(Long status, String Message) {
+public record ErrorSchema(int status, String Message) {
 }
